Add "Why should I care?" section and mention The Changelog.
I figured adding this section was notable enough to make a new
CHANGELOG entry as well.

The podcast is a bit long so I don't expect people new to this project
to actually listen to it right away but it would be good to extract
some of the good points made on the show into more fuel to the fire.