Some changes to sections and references.
Including sections with no notables changes by putting a single
"- Nothing." entry was a bad idea because for every single release this
created a ton of noise. My only gripe with this decision is that users
now have to assume that — for instance — the "Deprecated" section
was intentionally left out because it contained nothing, while it's quite
possible that they maintainers may not have deemed deprecations
worthy of a change log entry. That's a problem because it makes
expectations hard to manage.

Another change I've made is that I now refer to a "change log" (lower case, two words) instead of a "CHANGELOG" (uppercase, one word)
because I've released that the file is named "CHANGELOG" but the
purpose of the file matters more than its name. Moreover the previous
spelling made for awkward sentences and style.