Use iframe pointing to CHANGELOG instead of image.
The static image was a stop-gap to give people a visual rendition of a
raw CHANGELOG file. I find this iframe solution much more satisfying
because it allows people to scroll around the CHANGELOG file.

It's also a LOT easier to maintain since... I don't have to update the
image ever again. Wee!