Fixes german-language version of document
I rewrote some parts of the german-language version of the document to
make it sound more natural (to me). This includes changing some
expressions to be a somehwat more colloquial, and changing some parts to
not use the second person form (to avoid awkwardness with german
polite/impolite form of 'you').